requirements.txt
# ============================================================================
# requirements.txt — Portrait Sketch Pipeline
# Triton 24.01 base image / Python 3.10 / CUDA 12.1
# ============================================================================

numpy==1.26.4
scipy==1.11.4
sympy==1.14.0

pillow==10.2.0
opencv-python-headless==4.9.0.80
scikit-image==0.22.0
imageio==2.33.1
tifffile==2023.7.10
lazy_loader==0.3

diffusers==0.25.1
transformers==4.39.3
accelerate==0.28.0
safetensors==0.7.0
peft==0.5.0

huggingface-hub==0.25.2
tokenizers==0.15.2
filelock==3.19.1
fsspec==2025.9.0
regex==2025.11.3
packaging==25.0

controlnet-aux==0.0.7
timm==1.0.22
einops==0.8.1

facenet-pytorch==2.6.0
insightface==0.7.3
easydict==1.13

# ── ONNX ─────────────────────────────────────────────────────────────────────
onnx==1.20.0
onnxruntime-gpu==1.17.1

# ── rembg runtime deps (rembg itself installed --no-deps in Dockerfile) ───────
pooch==1.8.2                # rembg uses this to locate model files
PyMatting==1.1.14

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm==4.67.1
psutil==7.1.3
PyYAML==6.0.3
requests==2.28.1
urllib3==1.26.13
charset-normalizer==2.1.1
idna==3.4
certifi==2022.12.7
click==8.3.1
typing_extensions==4.15.0
pydantic==2.12.5
pydantic_core==2.41.5       # required by pydantic v2
annotated-types==0.7.0      # required by pydantic v2
Jinja2==3.1.6               # required by torch/diffusers
MarkupSafe==2.1.5           # required by Jinja2
networkx==3.3               # required by torch
jsonschema==4.26.0
